chore: Add function to generate Pascal's triangle
The code changes include adding a function `pascal_triangle` that generates Pascal's triangle up to a given number of rows. The function takes an integer `n` as input and returns a list of lists representing Pascal's triangle.

This commit message suggests using the `chore` type to indicate that this is a non-functional change that improves the code structure or development process. The message also provides a brief description of the added function and its purpose.